refactor: dto 폴더 presentaion 으로 이동
#60
