+++ com.bulbas23r.client.delivery/src/main/java/com/bulbas23r/client/delivery/application/service/DeliveryService.java
@@ -1,10 +1,11 @@
 package com.bulbas23r.client.delivery.application.service;
 
-import com.bulbas23r.client.delivery.application.dto.DeliveryCreateRequestDto;
-import com.bulbas23r.client.delivery.application.dto.DeliveryResponseDto;
-import com.bulbas23r.client.delivery.application.dto.DeliverySearchRequestDto;
-import com.bulbas23r.client.delivery.application.dto.DeliveryUpdateRequestDto;
+import com.bulbas23r.client.delivery.presentation.dto.request.DeliveryCreateRequestDto;
+import com.bulbas23r.client.delivery.presentation.dto.response.DeliveryResponseDto;
+import com.bulbas23r.client.delivery.presentation.dto.request.DeliverySearchRequestDto;
+import com.bulbas23r.client.delivery.presentation.dto.request.DeliveryUpdateRequestDto;
 import com.bulbas23r.client.delivery.domain.model.DeliveryStatus;
+import common.event.CreateOrderEventDto;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,4 +18,5 @@ public interface DeliveryService {
     DeliveryResponseDto deleteDelivery(UUID deliveryId);
     Page<DeliveryResponseDto> searchDelivery(DeliverySearchRequestDto requestDto);
     void changeStatus(UUID deliveryId, DeliveryStatus status);
+    void createDeliveryByOrder(CreateOrderEventDto eventDto);
 }
