+++ com.bulbas23r.client.delivery/src/main/java/com/bulbas23r/client/delivery/presentation/dto/request/DeliveryRouteArriveRequestDto.java
@@ -1,4 +1,4 @@
-package com.bulbas23r.client.delivery.application.dto;
+package com.bulbas23r.client.delivery.presentation.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
@@ -7,10 +7,7 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class DeliveryRouteDepartRequestDto {
-
-    @NotNull
-    private UUID deliveryManagerId;
+public class DeliveryRouteArriveRequestDto {
 
     @NotNull
     private UUID deliveryId;
@@ -21,5 +18,8 @@ public class DeliveryRouteDepartRequestDto {
     @NotNull
     private UUID arrivalHubId;
 
+    private Integer distance;
+
+    private Integer duration;
 
 }
